+++ gpu_box_benchmark/benchmark_dockerfile_wrappers/ai_benchmark.py
@@ -15,7 +15,16 @@ from gpu_box_benchmark.numeric_benchmark_result import BenchmarkResult, Numerica
 
 LOGGER = logging.getLogger(__name__)
 
-_AI_BENCHMARK_VERSION = "0.1.0"
+_AI_BENCHMARK_VERSION = "0.2.0"
+"""
+# Version History
+
+## 0.2.0 - (2026-02-03)
+* Switched to base image w/CUDA 11.4.3 to support Kepler era cards.
+
+## 0.1.0 - (2026-01-20)
+* First version 
+"""
 
 
 def _parse_ai_score(container_outputs: ContainerOutputs) -> float:
@@ -66,7 +75,8 @@ def create_ai_benchmark_executor(  # pylin
             critical_result_key=NumericalResultKey.forced_multi_gpu_sum,
             numerical_results=docker_wrapper.benchmark_dockerfile(
                 dockerfile_path=AI_BENCHMARK_DOCKERFILE,
-                tag_prefix=benchmark_name.value,
+                benchmark_name=benchmark_name.value,
+                benchmark_version=_AI_BENCHMARK_VERSION,
                 gpus=gpus,
                 create_runtime_env_vars=lambda runtime_gpus: [],
                 multi_gpu_native=multi_gpu_native,