+++ gpu_box_benchmark/benchmark_dockerfile_wrappers/content_aware_timelapse.py
@@ -15,7 +15,17 @@ from gpu_box_benchmark.numeric_benchmark_result import BenchmarkResult, Numerica
 
 LOGGER = logging.getLogger(__name__)
 
-_CONTENT_AWARE_TIMELAPSE_BENCHMARK_VERSION = "0.1.0"
+_CONTENT_AWARE_TIMELAPSE_BENCHMARK_VERSION = "0.2.0"
+"""
+# Version History
+
+## 0.2.0 - (2026-02-03)
+* Switched to base image w/CUDA 11.4.3 to support Kepler era cards.
+* Bumped underlying CAT version to v0.17.5
+
+## 0.1.0 - (2026-01-20)
+* First version 
+"""
 
 
 def _parse_fps(container_outputs: ContainerOutputs) -> float:
@@ -72,7 +82,8 @@ def create_content_aware_timelapse_executor(  # pylin
             critical_result_key=NumericalResultKey.native_multi_gpu_result,
             numerical_results=docker_wrapper.benchmark_dockerfile(
                 dockerfile_path=CONTENT_AWARE_TIMELAPSE_DOCKERFILE,
-                tag_prefix=benchmark_name.value,
+                benchmark_name=benchmark_name.value,
+                benchmark_version=_CONTENT_AWARE_TIMELAPSE_BENCHMARK_VERSION,
                 gpus=gpus,
                 create_runtime_env_vars=lambda runtime_gpus: envs,
                 multi_gpu_native=multi_gpu_native,