+++ gpu_box_benchmark/benchmark_dockerfile_wrappers/nvidia_deep_learning_examples.py
@@ -20,7 +20,16 @@ from gpu_box_benchmark.numeric_benchmark_result import BenchmarkResult, Numerica
 
 LOGGER = logging.getLogger(__name__)
 
-_RESNET50_BENCHMARK_VERSION = "0.1.0"
+_RESNET50_BENCHMARK_VERSION = "0.2.0"
+"""
+# Version History
+
+## 0.2.0 - (2026-02-03)
+* Switched to base image w/CUDA 11.4.3 to support Kepler era cards.
+
+## 0.1.0 - (2026-01-20)
+* First version 
+"""
 
 
 class _ResNet50Params(NamedTuple):
@@ -29,7 +38,6 @@ class _ResNet50Params(NamedTuple):
     """
 
     batch_size: int
-    amp_enabled: bool
 
     mode_training: bool
     """
@@ -75,17 +83,11 @@ def create_resnet50_executor(
     """
 
     name_to_parameters = {
-        BenchmarkName.resnet50_train_batch_1_amp: _ResNet50Params(
-            mode_training=True, batch_size=1, amp_enabled=True
-        ),
-        BenchmarkName.resnet50_train_batch_64_amp: _ResNet50Params(
-            mode_training=True, batch_size=64, amp_enabled=True
-        ),
-        BenchmarkName.resnet50_infer_batch_1_amp: _ResNet50Params(
-            mode_training=False, batch_size=1, amp_enabled=True
-        ),
-        BenchmarkName.resnet50_infer_batch_256_amp: _ResNet50Params(
-            mode_training=False, batch_size=256, amp_enabled=True
+        BenchmarkName.resnet50_train_batch_1: _ResNet50Params(mode_training=True, batch_size=1),
+        BenchmarkName.resnet50_train_batch_64: _ResNet50Params(mode_training=True, batch_size=64),
+        BenchmarkName.resnet50_infer_batch_1: _ResNet50Params(mode_training=False, batch_size=1),
+        BenchmarkName.resnet50_infer_batch_256: _ResNet50Params(
+            mode_training=False, batch_size=256
         ),
     }
 
@@ -113,7 +115,8 @@ def create_resnet50_executor(
             critical_result_key=NumericalResultKey.native_multi_gpu_result,
             numerical_results=docker_wrapper.benchmark_dockerfile(
                 dockerfile_path=RESNET50_DOCKERFILE,
-                tag_prefix=benchmark_name.value,
+                benchmark_name=benchmark_name.value,
+                benchmark_version=_RESNET50_BENCHMARK_VERSION,
                 gpus=gpus,
                 create_runtime_env_vars=lambda runtime_gpus: [
                     ("BATCH_SIZE", str(resnet_parameters.batch_size)),