+++ gpu_box_benchmark/docker_wrapper.py
@@ -27,12 +27,6 @@ output directly to a file. This is the path the containers write to internally, 
 that volume mount. 
 """
 
-_SESSION_ID_LABEL_KEY = "gpu_box_benchmark_session"
-_SESSION_ID_BUILD_ARG = "SESSION_ID"
-"""
-Images are labeled so smart cleanup can occur. 
-"""
-
 
 class ContainerOutputs(NamedTuple):
     """
@@ -124,6 +118,7 @@ def _create_gpu_container(
 def _wait_get_logs(container: Container) -> Optional[str]:
     """
     Runs a `.wait()` on the input container and then pulls the stdout/stderr logs.
+    Removes the container as well.
     :param container: To interact with.
     :return: The logs as a single string.
     """
@@ -154,7 +149,7 @@ def _wait_get_logs(container: Container) -> Optional[str]:
     return logs
 
 
-def _run_image_on_gpus(
+def _run_image_on_gpus(  # pylint: disable=too-many-positional-arguments
     client: DockerClient,
     image: Image,
     gpus: Tuple[GPUIdentity, ...],
@@ -188,7 +183,9 @@ def _run_image_on_gpus(
     )
 
 
-def _build_image(client: DockerClient, dockerfile_path: Path, tag: str, session_id: str) -> Image:
+def _build_image(
+    client: DockerClient, dockerfile_path: Path, tag: str, docker_cleanup: bool
+) -> Image:
     """
     Canonical wrapper to build images.
 
@@ -196,10 +193,18 @@ def _build_image(client: DockerClient, dockerfile_path: Path, tag: str, session_
     :param dockerfile_path: Path to the dockerfile to build. The build context will be the parent
     of this file.
     :param tag: Tag for the image.
+    :param docker_cleanup: If True, efforts are made to leave little behind related to docker.
     :return: The built image.
     """
 
     LOGGER.debug("Building Image")
+
+    if not docker_cleanup:
+        try:
+            LOGGER.debug(f"Checking for existing image {tag}...")
+            return client.images.get(tag)
+        except docker.errors.ImageNotFound:
+            LOGGER.debug("Image not found locally. Proceeding to build.")
 
     build_directory = dockerfile_path.parent
 
@@ -208,14 +213,10 @@ def _build_image(client: DockerClient, dockerfile_path: Path, tag: str, session_
             path=str(build_directory),
             dockerfile=str(dockerfile_path),
             tag=tag,
-            # All dockerfiles accept this build arg.
-            buildargs={_SESSION_ID_BUILD_ARG: session_id},
-            # These labels are used for cleanup.
-            labels={_SESSION_ID_LABEL_KEY: session_id},
-            nocache=True,
+            nocache=docker_cleanup,
             rm=True,
             forcerm=True,
-            pull=False,
+            pull=docker_cleanup,
         )
 
         return image
@@ -302,32 +303,16 @@ def _list_prunable_images(client: docker.DockerClient) -> List[Image]:
     return [image for image in client.images.list() if image.id not in used_image_ids]
 
 
-def _docker_image_cleanup(
-    client: DockerClient, run_image: Image, session_id: str, keep_images: List[Image]
-) -> None:
+def _docker_image_cleanup(client: DockerClient, run_image: Image, keep_images: List[Image]) -> None:
     """
     Removes dangling images and containers created throughout the benchmarking process.
     :param client: For interacting with Docker.
     :param run_image: Image that was run.
-    :param session_id: Label
     :param keep_images: List of images to not touch.
     :return: None
     """
 
     client.images.remove(run_image.id, force=True)
-
-    session_filter = {"label": f"{_SESSION_ID_LABEL_KEY}={session_id}"}
-
-    for container in client.containers.list(all=True, filters=session_filter):
-        LOGGER.warning(f"Removing container {container.id}")
-        container.remove(force=True)
-
-    prune_report = client.images.prune(filters=session_filter)
-    reclaimed = prune_report.get("SpaceReclaimed", 0)
-    LOGGER.log(
-        msg=f"Pruned session images. Reclaimed: {reclaimed} bytes",
-        level=logging.WARNING if reclaimed > 0 else logging.DEBUG,
-    )
 
     parent_id = run_image.attrs.get("Parent") if run_image else None
 
@@ -358,7 +343,8 @@ def _docker_image_cleanup(
 
 def benchmark_dockerfile(  # pylint: disable=too-many-positional-arguments,too-many-locals
     dockerfile_path: Path,
-    tag_prefix: str,
+    benchmark_name: str,
+    benchmark_version: str,
     gpus: Tuple[GPUIdentity, ...],
     create_runtime_env_vars: CreateRuntimeEnvironmentVariables,
     outputs_to_result: OutputsToResult,
@@ -380,7 +366,8 @@ def benchmark_dockerfile(  # pylint: disable=too-many-positional-arguments,too-m
     This shouldn't impact GPU scores and will likely be improved in the future.
 
     :param dockerfile_path: Path to the dockerfile to run.
-    :param tag_prefix: Tag prefix for the image. The actual image will be uniquely tagged.
+    :param benchmark_name: Passed to image tagging process.
+    :param benchmark_version: Passed to image tagging process.
     :param gpus: GPUs to run on.
     :param create_runtime_env_vars: Callable to create runtime environment variables.
     :param outputs_to_result: Converts the outputs from the container (logs and optional output
@@ -401,14 +388,11 @@ def benchmark_dockerfile(  # pylint: disable=too-many-positional-arguments,too-m
     # touch these because they might be used for something outside the scope of our app.
     existed_before_run = _list_prunable_images(client=client)
 
-    # Passed to the build process and used in the cleanup process.
-    session_id = str(uuid.uuid4().hex)
-
     image: Image = _build_image(
         client=client,
         dockerfile_path=dockerfile_path,
-        tag="_".join([tag_prefix, session_id]),
-        session_id=session_id,
+        tag="_".join([benchmark_name, benchmark_version]),
+        docker_cleanup=docker_cleanup,
     )
 
     # If there's only a single GPU attached we can skip a lot of work.
@@ -497,7 +481,6 @@ def benchmark_dockerfile(  # pylint: disable=too-many-positional-arguments,too-m
             _docker_image_cleanup(
                 client=client,
                 run_image=image,
-                session_id=session_id,
                 keep_images=existed_before_run,
             )